版本时间:2015.06.08

+++ rd/trunk/Server/ServerCode/hawkjava/GameServer/src/com/hawk/game/entity/MonsterEntity.java
@@ -125,8 +125,8 @@ public class MonsterEntity extends HawkDBEntity {
 		return level;
 	}
 
-	public void setLevel(short level) {
-		this.level = level;
+	public void setLevel(int level) {
+		this.level = (short)level;
 	}
 
 	public int getExp() {
